+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### Added
+
+- AMG Dosing Weight
+  - BeeCalc on intranet uses just Actual or AdjBW (using 0.4, if >125% ibw) - this doesn't match protocol. Changed to match 2018 protocol update.
+- Body surface area calculation
+
+
 ## [Unreleased] - v1.0.2-beta.1
 
 ### Added